Remove recommendation of NASA license
The NASA license is AWFUL, and even many in NASA want to eliminate it.  For example, a straightforward reading of the NASA license requires that ONLY the original contributor be able to make changes to it.  Thus, if a DoD employee wants to add code that a dead DoE employee created to some NASA software, they CANNOT, because the DoD employee is not the "original author".  I suspect that wasn't intended, but after almost 10 years of pleading with NASA to clarify this, NASA has failed to do so.  This license has been explicitly rejected by the FSF.  In general, we need to STOP using government vanity licenses, and use industry-standard licenses (ranging from CC0 through common OSI licenses) for OSS works.